requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
cron_descriptor==2.0.6
dj-database-url==3.0.1
Django==4.2.7
django-celery-beat==2.8.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.12.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dnspython==2.8.0
gunicorn==23.0.0
idna==3.10
kombu==5.5.4
mongoengine==0.29.1
packaging==25.0
paystack==1.5.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.8.0
pymongo==4.15.3
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-env==1.0.0
pytz==2025.2
redis==7.0.1
requests==2.32.5
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
unicorn==2.1.4
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
whitenoise==6.11.0
